movi logica de jornadas a las correspondientes strategies
